+++ .env
@@ -1,0 +1,1 @@
+REACT_APP_API_BASE_URL=https://example.com
